api/requirements.txt
flask
flask-cors
python-bitcoinrpc